Move message processing so we can add metrics to the functions
